Summary
The emerging trend of Open Data (OD) holds significant importance as it encompasses diverse datasets. The expansive nature of OD is a powerful tool with immense potential to empower data publishers, analysts, citizens and stakeholders by making data more discoverable, available and usable. Linked Open Data (LOD) is a process that enhances transparency and innovation in government functions.
While OD offers many advantages, it also introduces several challenges, which are the focus of this thesis. Heterogeneity, which stems from the diverse range of data formats. Data quality, ambiguity and completeness, highlighting the importance of establishing standards and performance indicators for data analysis, semantic annotation, repair, augmentation, and visualization. Discoverability and understandability is highlighted, which implies the integration of semantic links and the fluid flow of data between stakeholders.
This doctoral thesis is dedicated to surmounting obstacles associated with OD, concurrently presenting remedies to enhance practical interventions in this domain. We propose an efficient and integrated semi-automated framework designed to inter-actively aid OD publishers in the integration and publication of LOD.  Our objective is to enhance the usability, discoverability, and understandability of OD. We introduce an all-in-one automated framework capable of addressing data quality issues and supporting annotation systems seamlessly interpreted into any OD portal. It is intended to expedite the identification of OD problems for data analysts, reducing the time required for problem detection. Concurrently, we are developing a user-friendly application to facilitate the merging of semantically annotated related datasets. The ultimate objectives are to improve data transparency and potentially lead to innovative solutions for the advancement of the community. Finally, we conduct an evaluation of the capabilities, usability, and efficacy of these contributions.